feat: ViewModel 구조 개선
- AppState, TaskInteractors 삭제
- TaskService 추가
